Fix bugs and clean up code related to ANY
Realized my tests that were expecting failure were broken.  If the
expect didn't fail, the catch wasn't reached.  Changed them to correctly
expect the exception to be thrown.

The failed tests were hiding the fact that toHaveBeenCalledWith was
broken and wouldn't catch a mismatched number of arguments.  Added a
length check to the assert to fix those bugs.

Also realized that the spyOn-test spy was accumulating all the calls,
and wasn't getting reset between tests, so I moved the video object
creation into the beforeEach to fix that problem.

Added a toString to the ANY class, so that when failures are reported,
they print 'ANY' instead of '{}'.